backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# File parsing
pymupdf==1.24.5
python-pptx==0.6.23
python-docx==1.1.2

# File uploads
python-multipart==0.0.9

# Environment variables
python-dotenv==1.0.1

# Validation
pydantic==2.9.0
typing-extensions==4.14.0

# LLM
google-generativeai==0.7.2

# Embedding + Vector Store
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy>=1.26,<2.0

# ── NEW: Authentication ─────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT encoding/decoding
httpx==0.28.0                     # Async HTTP for Google OAuth token exchange

# ── NEW: Rate Limiting ──────────────────────────────────────────────────────
slowapi==0.1.9                      # Rate limiting middleware for FastAPI

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7

# Utilities
aiofiles==23.2.1